Refactor course-study
