+++ frontend/package.json
@@ -167,6 +167,7 @@
     "ack-angular-webcam": "^1.9.3",
     "chart.js": "^2.7.0",
     "core-js": "^2.4.1",
+    "detectrtc": "^1.3.6",
     "dotenv": "^4.0.0",
     "file-saver": "^1.3.3",
     "font-awesome": "^4.7.0",
@@ -186,6 +187,7 @@
     "recordrtc": "^5.4.6",
     "reflect-metadata": "^0.1.10",
     "rxjs": "^5.4.2",
+    "screenfull": "^3.3.2",
     "socket.io-client": "^2.1.0",
     "systemjs": "0.20.14",
     "underscore": "^1.8.3",
