+++ frontend/src/client/app/app.module.ts
@@ -31,8 +31,8 @@ import { TranslateService } from '@ngx-translate/core';
         HttpModule,
         AppRoutingModule,
         AuthModule,
-        LMSModule,
         HomeModule.forRoot(),
+        LMSModule,
         DashboardModule,
         SettingModule,
         AccountModule,
