+++ frontend/src/client/app/cms/course/course-unit-preview-dialog/course-unit-preview-dialog.component.html
@@ -1,4 +1,3 @@
-<form novalidate (ngSubmit)="save()" #f="ngForm" autocomplete="off">
   <p-dialog header="{{nameUnit}}" [(visible)]="display" modal="true" [responsive]="true" width="1000" appendTo="body">
     <div class="spinner" [hidden]="!loading"></div>
     <p-scrollPanel [style]="{width: '100%'}">
@@ -16,4 +15,3 @@
       <button type="button" pButton icon="fa-close" (click)="hide()" label="{{'Close'|translate}}"></button>
     </p-footer>
   </p-dialog>
-</form>
