+++ frontend/tools/config/project.config.ts
@@ -92,6 +92,10 @@ export class ProjectConfig extends SeedConfig {
                 path: 'node_modules/detectrtc/DetectRTC.min.js'
             },
             {
+                name: 'screenfull',
+                path: 'node_modules/screenfull/dist/screenfull.js'
+            },
+            {
                 name: 'socket.io-client',
                 path: 'node_modules/socket.io-client/dist/socket.io.js'
             },
